Can now provide new users with a starting date.
Before, users were given a start date that defaulted to the day they
were created. Now, the new user form allows you to control the start
date so that you can add new users that may have already started, or
will be starting in the future.

This change allows administrators to add all of the users when they
start using the application, rather than having some fixture script that
created all of the veteran users.